+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added documentation for `create-post` command.
 - Mark portable actor as reachable if at least one inbox is reachable.
 - Don't retry delivery if status is `410 Gone`.
+- Prioritize primary recipients when delivering posts.
 
 ### Fixed